// src/type.ts
export type characterList =
  | 'milk-cow1'
  | 'milk-cow2'
  | 'quokka'
  | 'elephant1'
  | 'elephant2'
  | 'rabbit1'
  | 'rabbit2'
  | 'frog'
  | 'duck1'
  | 'duck2'
  | 'pig'
  | 'puppy'
  | 'bear1'
  | 'bear2'
  | 'giraffe1'
  | 'giraffe2';

type personType =
  | 'INTJ'
  | 'ISTJ'
  | 'ENFJ'
  | 'ISFP'
  | 'ISTP'
  | 'ENTP'
  | 'ESTP'
  | 'ENFP'
  | 'ISFJ'
  | 'INFJ'
  | 'ESFJ'
  | 'ESFP'
  | 'INFP'
  | 'INTP'
  | 'ESTJ'
  | 'ENTJ';